renamed commandgroups to autocommands
