+++ src/org/usfirst/frc/team1250/robot/AutoGroups/Auto_RightPos.java
@@ -1,8 +1,8 @@
 package org.usfirst.frc.team1250.robot.AutoGroups;
 
 import org.usfirst.frc.team1250.robot.Robot;
-import org.usfirst.frc.team1250.robot.commandGroups.CmdG_ScaleWhip;
-import org.usfirst.frc.team1250.robot.commandGroups.CmdG_StriaghtScale;
+import org.usfirst.frc.team1250.robot.AutoCommands.CmdG_ScaleWhip;
+import org.usfirst.frc.team1250.robot.AutoCommands.CmdG_StriaghtScale;
 import org.usfirst.frc.team1250.robot.drive.Cmd_AutoDrive;
 import org.usfirst.frc.team1250.robot.drive.Cmd_AutoTurn;
 import org.usfirst.frc.team1250.robot.elevator.Cmd_EleErectTower;
