+++ src/org/usfirst/frc/team1250/robot/OI.java
@@ -10,14 +10,14 @@ package org.usfirst.frc.team1250.robot;
 import edu.wpi.first.wpilibj.Joystick;
 import edu.wpi.first.wpilibj.buttons.*;
 
-import org.usfirst.frc.team1250.robot.commandGroups.CmdG_AutoLeft;
-import org.usfirst.frc.team1250.robot.commandGroups.CmdG_Droper;
-import org.usfirst.frc.team1250.robot.commandGroups.CmdG_FullCollect;
-import org.usfirst.frc.team1250.robot.commandGroups.CmdG_AutonSwitch;
-import org.usfirst.frc.team1250.robot.commandGroups.CmdG_ScaleWhip;
-import org.usfirst.frc.team1250.robot.commandGroups.CmdG_StriaghtScale;
-import org.usfirst.frc.team1250.robot.commandGroups.CmdG_TimedCollect;
 import org.usfirst.frc.team1250.robot.commandIntake.Cmd_IntakeSpitFast;
+import org.usfirst.frc.team1250.robot.AutoCommands.CmdG_AutoLeft;
+import org.usfirst.frc.team1250.robot.AutoCommands.CmdG_AutonSwitch;
+import org.usfirst.frc.team1250.robot.AutoCommands.CmdG_Droper;
+import org.usfirst.frc.team1250.robot.AutoCommands.CmdG_FullCollect;
+import org.usfirst.frc.team1250.robot.AutoCommands.CmdG_ScaleWhip;
+import org.usfirst.frc.team1250.robot.AutoCommands.CmdG_StriaghtScale;
+import org.usfirst.frc.team1250.robot.AutoCommands.CmdG_TimedCollect;
 import org.usfirst.frc.team1250.robot.commandIntake.Cmd_IntakeSpit;
 import org.usfirst.frc.team1250.robot.intake.Cmd_IntakeCollectTimed;
 import org.usfirst.frc.team1250.robot.commands.*;
